+++ lib/model/event.rb
